Retire HAL for STM32F4 / F7 (#20153)
