+++ Marlin/src/HAL/shared/backtrace/unwmemaccess.cpp
@@ -74,17 +74,6 @@
   #define START_FLASH_ADDR  0x08000000
   #define END_FLASH_ADDR    0x08080000
 
-#elif MB(THE_BORG)
-
-  // For STM32F765 in BORG
-  //  SRAM  (0x20000000 - 0x20080000) (512kb)
-  //  FLASH (0x08000000 - 0x08100000) (1024kb)
-  //
-  #define START_SRAM_ADDR   0x20000000
-  #define END_SRAM_ADDR     0x20080000
-  #define START_FLASH_ADDR  0x08000000
-  #define END_FLASH_ADDR    0x08100000
-
 #elif MB(REMRAM_V1, NUCLEO_F767ZI)
 
   // For STM32F765VI in RemRam v1
