+++ platformio.ini
@@ -863,22 +863,13 @@ lib_deps          = ${common_stm32f1.lib_deps}
   USBComposite for STM32F1@0.91
 
 #
-# STM32F4 with STM32GENERIC
+# REMRAM_V1
 #
-[env:STM32F4]
+[env:REMRAM_V1]
 platform      = ${common_stm32.platform}
-board         = disco_f407vg
-build_flags   = ${common.build_flags} -DUSE_STM32GENERIC -DSTM32GENERIC -DSTM32F4 -DMENU_USB_SERIAL -DMENU_SERIAL=SerialUSB -DHAL_IWDG_MODULE_ENABLED
-src_filter    = ${common.default_src_filter} +<src/HAL/STM32_F4_F7> -<src/HAL/STM32_F4_F7/STM32F7>
-
-#
-# STM32F7 with STM32GENERIC
-#
-[env:STM32F7]
-platform      = ${common_stm32.platform}
+extends       = common_stm32
 board         = remram_v1
-build_flags   = ${common.build_flags} -DUSE_STM32GENERIC -DSTM32GENERIC -DSTM32F7 -DMENU_USB_SERIAL -DMENU_SERIAL=SerialUSB -DHAL_IWDG_MODULE_ENABLED
-src_filter    = ${common.default_src_filter} +<src/HAL/STM32_F4_F7> -<src/HAL/STM32_F4_F7/STM32F4>
+build_flags   = ${common_stm32.build_flags}
 
 #
 # ST NUCLEO-F767ZI Development Board
